backend/requirements.txt
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
Flask-Limiter==3.12
pymongo==4.11.3
qrcode[pil]==8.0
Pillow==11.1.0
Werkzeug==3.1.3
python-dotenv==1.1.0
bleach==6.2.0
marshmallow==3.25.1
